// src-ui/core/ipc/generated/domain/TerminalTarget.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * 终端开到哪
 */
export type TerminalTarget = { "kind": "local" } | { "kind": "server", server_id: string, } | { "kind": "bot", bot_id: string, host_dir: boolean, } | { "kind": "app_instance", instance_id: string, };